Feat diner

+++ src/controllers/diners.controller.js
@@ -26,6 +26,9 @@ export class DinersController {
   // 식당 등록
   createDiner = async (req, res, next) => {
     try {
+      const { type: userType, adminId } = res.locals.user;
+      if (userType !== 'ADMIN')
+        return res.status(401).json({ message: '권한이 없습니다.' });
       const {
         name,
         type,
@@ -49,7 +52,13 @@ export class DinersController {
         return res
           .status(400)
           .json({ message: '영업 시간 형식이 적절하지 않습니다.' });
+      const existDiner = await this.dinersService.getDiner({ adminId });
+      if (existDiner)
+        return res
+          .status(400)
+          .json({ message: '이미 등록하신 매장이 존재합니다.' });
       const newDiner = await this.dinersService.createDiner(
+        adminId,
         name,
         type,
         address,
@@ -70,6 +79,15 @@ export class DinersController {
     res.json({ diners });
   };
 
+  // 식당 키워드 검색
+  searchDiners = async (req, res) => {
+    const { key } = req.query;
+    if (!key)
+      return res.status(400).json({ message: '검색어를 입력해주세요.' });
+    const diners = await this.dinersService.searchDiners(key);
+    res.json({ diners });
+  };
+
   // 특정 식당 찾기
   findDiner = async (req, res, next) => {
     const dinerId = +req.params.dinerId;
@@ -77,7 +95,7 @@ export class DinersController {
       return res
         .status(404)
         .json({ message: '해당 매장이 존재하지 않습니다.' });
-    const diner = await this.dinersService.getDiner(dinerId);
+    const diner = await this.dinersService.getDiner({ dinerId });
     if (!diner)
       return res
         .status(404)
@@ -92,6 +110,9 @@ export class DinersController {
   // 식당 정보 수정
   updateDiner = async (req, res, next) => {
     try {
+      const { type: userType, adminId } = res.locals.user;
+      if (userType !== 'ADMIN' || adminId !== res.locals.diner.adminId)
+        return res.status(401).json({ message: '권한이 없습니다.' });
       const {
         name,
         type,
@@ -135,6 +156,10 @@ export class DinersController {
 
   // 식당 삭제
   deleteDiner = async (req, res, next) => {
+    // 인증 수정
+    const { type: userType, adminId } = res.locals.user;
+    if (userType !== 'ADMIN' || adminId !== res.locals.diner.adminId)
+      return res.status(401).json({ message: '권한이 없습니다.' });
     await this.dinersService.deleteDiner(res.locals.diner.dinerId);
     res.status(204).json();
   };
